Rename files from 'reconciliation' to 'sfc internal'
